Refactoring (main) - 41
