+++ controllers/contacts/addContact.js
@@ -1,0 +1,74 @@
+const Joi = require('joi')
+
+const contactsOperations = require("../../models/contacts")
+const { lineBreak } = require("../../service");
+
+
+//-----------------------------------------------------------------------------
+//* ++++++++++++++++++++++ Схема ВАЛИДАЦИИ Joi +++++++++++++++++++++++++
+const contactSchemaPostPut = Joi.object({
+    name: Joi.string()
+        // .alphanum()
+        .min(3)
+        .max(30)
+        .required(),
+
+    email: Joi.string()
+        .email({ minDomainSegments: 2, tlds: { allow: ['com', 'net', 'ua', 'org',] } })
+        .required(),
+
+    phone: Joi.string()
+        // .alphanum()
+        .min(5)
+        .max(14)
+        .required(),
+});
+//* _______________________ Схема ВАЛИДАЦИИ Joi _______________________
+
+
+
+const addContact = async (req, res, next) => {
+    try {
+        //! ===========================console============================
+        console.log("START-->POST".yellow); //!
+        lineBreak();
+        //! ==============================================================
+
+        //* +++++++++++++++++++++++ ВАЛИДАЦИЯ Joi +++++++++++++++++++++++++++++
+        const validationResult = contactSchemaPostPut.validate(req.body);
+
+        if (validationResult.error) {
+            //! ===========================console============================
+            console.log("Ошибка ВАЛИДАЦИИ:".bgRed.black);
+            console.log("");
+            console.log(validationResult.error);
+            lineBreak();
+            console.log("END-->POST".yellow); //!
+            //! ==============================================================
+            //! 1 - вариант
+            // return res.status(400).json({ "message": "missing required name field" });
+            //! 2 - вариант
+            // validationResult.error.status = 400
+            // throw validationResult.error
+            //! 3 - вариант
+            return res.status(400).json({ status: validationResult.error.details });
+        }
+        //* __________________________ ВАЛИДАЦИЯ Joi __________________________
+
+        const contact = await contactsOperations.addContact(req.body)
+
+        res.status(201).json({
+            status: "success",
+            code: 201,
+            data: {
+                result: contact
+            }
+        });
+
+    } catch (e) {
+        next(e);
+        // res.status(500).json({ error: e.message });
+    }
+}
+
+module.exports = addContact
