+++ controllers/contacts/index.js
@@ -1,7 +1,12 @@
 const getAllContacts = require("./getAllContacts")
 const getContactById = require("./getContactById")
+const addContact = require("./addContact")
+
+
+
 
 module.exports = {
     getAllContacts,
-    getContactById
+    getContactById,
+    addContact
 }
