+++ routes/api/contacts.js
@@ -67,49 +67,7 @@ router.get('/:contactId', ctrl.getContactById)
 
 //-----------------------------------------------------------------------------
 //! 3. Создание НОВОГО ПОЛЬЗОВАТЕЛЯ
-router.post("/", async (req, res, next) => {
-  try {
-    //! ===========================console============================
-    console.log("START-->POST".yellow); //!
-    lineBreak();
-    //! ==============================================================
-
-    //* +++++++++++++++++++++++ ВАЛИДАЦИЯ Joi +++++++++++++++++++++++++++++
-    const validationResult = contactSchemaPostPut.validate(req.body);
-
-    if (validationResult.error) {
-      //! ===========================console============================
-      console.log("Ошибка ВАЛИДАЦИИ:".bgRed.black);
-      console.log("");
-      console.log(validationResult.error);
-      lineBreak();
-      console.log("END-->POST".yellow); //!
-      //! ==============================================================
-      //! 1 - вариант
-      // return res.status(400).json({ "message": "missing required name field" });
-      //! 2 - вариант
-      // validationResult.error.status = 400
-      // throw validationResult.error
-      //! 3 - вариант
-      return res.status(400).json({ status: validationResult.error.details });
-    }
-    //* __________________________ ВАЛИДАЦИЯ Joi __________________________
-
-    const contact = await contactsOperations.addContact(req.body)
-
-    res.status(201).json({
-      status: "success",
-      code: 201,
-      data: {
-        result: contact
-      }
-    });
-
-  } catch (e) {
-    next(e);
-    // res.status(500).json({ error: e.message });
-  }
-});
+router.post("/", ctrl.addContact);
